Add Express backend with persistent Swiss tournament state
