+++ index.html
@@ -12,6 +12,9 @@
       <p class="tagline">Web recreation of the SwissChessLeaguev4.xlsm workbook</p>
     </header>
     <main id="app"></main>
+    <script>
+      window.APP_API_BASE = window.APP_API_BASE || 'http://localhost:4000/api';
+    </script>
     <script type="module" src="src/app.js"></script>
   </body>
 </html>
