+++ package.json
@@ -1,0 +1,22 @@
+{
+  "name": "swiss-chess-league",
+  "version": "1.0.0",
+  "type": "module",
+  "scripts": {
+    "start:server": "node server/index.js",
+    "dev:server": "nodemon server/index.js",
+    "migrate": "node server/scripts/migrate.js",
+    "seed": "node server/scripts/seed.js",
+    "reset": "node server/scripts/reset.js",
+    "start:client": "http-server . -c-1 -p 4173"
+  },
+  "dependencies": {
+    "better-sqlite3": "^8.6.0",
+    "cors": "^2.8.5",
+    "express": "^4.19.2"
+  },
+  "devDependencies": {
+    "http-server": "^14.1.1",
+    "nodemon": "^3.0.3"
+  }
+}
