+++ server/scripts/reset.js
@@ -1,0 +1,5 @@
+import { seedDatabase } from './seed.js';
+
+seedDatabase();
+
+console.log('Tournament reset to the initial seeded state.');
